+++ lib/flutter_duit.dart
@@ -1,7 +1,8 @@
 library flutter_duit;
 
 export "package:flutter_duit/src/duit_impl/index.dart";
-export "package:flutter_duit/src/ui/index.dart";
+export "package:flutter_duit/src/ui/index.dart"
+    show CustomUiElement, AttendedModel;
 export "package:flutter_duit/src/transport/transport_type.dart";
 export "package:flutter_duit/src/transport/options.dart";
 export "package:flutter_duit/src/kernel_api/index.dart";
@@ -10,7 +11,7 @@ export "package:flutter_duit/src/animations/index.dart"
 export "package:flutter_duit/src/utils/index.dart"
     hide
         GestureInterceptionLogic,
-        AttributeParser,
+        DefaultAttributeParser,
         ImageType,
         JsonUtils,
         JSONObject,