+++ lib/src/duit_impl/index.dart
@@ -1,4 +1,3 @@
 export 'driver.dart';
 export 'host.dart';
 export 'view_controller_change_listener.dart';
-export 'external_event_handler.dart';